recipe_site/requirements.txt
djangorestframework==3.14.0
Pillow==10.2.0
